+++ doc/Changelog.md
@@ -4,15 +4,19 @@
 - copy buttons
 - save addresses to wallet file
 - export addresses
-- cargo clippy (187 warnings)
+- mnemonic passphrase upper limit in settings
+
+# version 0.74.3
+- cargo clippy - 3 warnings
 
 # version 0.74.2
-- remove cargo clippy warnings >100
+- cargo clippy - 100 warnings
 - cargo crates updated
 
 # version 0.74.1
 - save gui window size
 - cleaning mess after rust formatter
+- cargo clippy - 187 warnings
 
 # version 0.74.0
 - beta proxy (not tested yet)